requirements.txt
accelerate==0.29.3
addict==2.4.0
aiofiles==23.2.1
aiohttp==3.9.3
aioresponses==0.7.6
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anthropic==0.21.3
anyio==4.3.0
appdirs==1.4.4
appnope==0.1.4
APScheduler==3.10.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
asyncpg==0.29.0
atlas-provider-sqlalchemy
attr==0.3.1
attrs==23.2.0
awscli==1.29.54
azure-cognitiveservices-vision-computervision==0.9.0
azure-common==1.1.28
azure-core==1.30.1
azure-identity==1.16.0
azure-storage-blob==12.19.1
Babel==2.14.0
backoff==2.2.1
beautifulsoup4==4.12.3
bleach==5.0.1
blinker==1.8.1
boilerpy3==1.0.7
boto==2.49.0
boto3==1.28.23
botocore==1.31.54
boxing==0.1.4
build==0.10.0
cachetools==5.3.3
cattrs==23.2.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
clip @ git+https://github.com/ultralytics/CLIP.git@e17416a36b45d040758327936a1ea150c13fe3d1
cloudpickle==3.0.0
cmake==3.29.2
colorama==0.4.4
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.2.1
cryptography==42.0.5
cssselect==1.2.0
cycler==0.12.1
Cython==3.0.0
dataclasses-json==0.6.5
datasets==2.19.0
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
Django==3.2.25
django-annoying==0.10.6
django-cors-headers==3.6.0
django-csp==3.7
django-debug-toolbar==3.2.1
django-environ==0.10.0
django-extensions==3.1.0
django-filter==2.4.0
django-model-utils==4.1.1
django-ranged-fileresponse==0.1.2
django-rq==2.5.1
django-storages==1.12.3
django-user-agents==0.4.0
djangorestframework==3.13.1
docker==6.1.3
docstring_parser==0.16
docutils==0.16
drf-dynamic-fields==0.3.0
drf-flex-fields==0.9.5
drf-generators==0.3.0
einops==0.8.0
exceptiongroup==1.2.0
executing==2.0.1
expiringdict==1.2.2
fastapi==0.110.1
fastjsonschema==2.19.1
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.13.3
flatbuffers==24.3.25
fonttools==4.50.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.3.1
ftfy==6.2.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.18.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
GPUtil==1.4.0
gradio_client==0.16.0
greenlet==3.0.3
guidance==0.1.13
h11==0.14.0
haystack-ai==2.0.0
haystack-bm25==1.0.2
htmlmin==0.1.12
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.22.2
humanfriendly==10.0
humansignal-drf-yasg==1.21.9
idna==3.6
ijson==3.2.3
imageio==2.34.1
importlib_metadata==7.1.0
inflection==0.5.1
iniconfig==2.0.0
instructor==1.2.5
interegular==0.3.3
ipykernel==6.29.4
ipython==8.23.0
ipywidgets==8.1.2
isodate==0.6.1
isoduration==20.11.0
jedi==0.19.1
jieba3k==0.35.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
json5==0.9.24
jsonformer==0.12.0
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.26.0
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
lark==1.1.9
launchdarkly-server-sdk==7.5.0
lazy-imports==0.3.1
lazy_loader==0.4
litellm==1.34.39
llvmlite==0.42.0
lm-format-enforcer==0.9.8
lmql==0.7.3
lockfile==0.12.2
lsprotocol==2023.0.1
lxml==5.2.1
lxml_html_clean==0.1.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
modelbit==0.37.0
monotonic==1.6
more-itertools==10.2.0
mpmath==1.3.0
msal==1.28.0
msal-extensions==1.1.0
msgpack==1.0.8
msrest==0.7.1
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
newspaper3k==0.2.8
ninja==1.11.1.1
nltk==3.6.7
notebook==7.1.2
notebook_shim==0.2.4
numba==0.59.1
numpy==1.25.2
oauthlib==3.2.2
ollama-haystack==0.0.5
onnxruntime==1.15.1
openai==1.16.1
opencv-python==4.8.0.76
opencv-python-headless==4.9.0.80
ordered-set==4.0.2
outlines==0.0.41
overrides==7.7.0
packaging==24.0
pandas==2.2.1
pandocfilters==1.5.1
parso==0.8.3
pendulum==3.0.0
pexpect==4.9.0
pgvector==0.2.5
piexif==1.1.3
pillow==10.3.0
pkginfo==1.10.0
platformdirs==4.2.0
playwright==1.43.0
pluggy==1.5.0
portalocker==2.8.2
posthog==3.5.0
prettytable==3.10.0
prometheus-fastapi-instrumentator==6.0.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
psycopg2-binary==2.9.6
ptyprocess==0.7.0
PuLP==2.8.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==16.0.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybase64==1.3.2
pyclipper==1.3.0.post5
pycocotools==2.0.7
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.9.0
pydot==2.0.0
pyee==11.1.0
pyformlang==1.0.9
pygls==1.3.1
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.2
pyproject_hooks==1.1.0
pyquery==2.0.0
pyRFC3339==1.1
pyrsistent==0.20.0
pytest==8.2.0
pytest-asyncio==0.21.1
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.4
python-multipart==0.0.9
pytz==2022.7.1
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
redis==3.5.3
referencing==0.34.0
regex==2023.12.25
replicate==0.25.2
requests==2.31.0
requests-file==2.0.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rf-groundingdino==0.2.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich
rpds-py==0.18.0
rq==1.10.1
rsa==4.7.2
ruff==0.4.2
ruff-lsp==0.0.53
rules==2.2
s3transfer==0.6.2
safetensors==0.4.3
scikit-image==0.23.2
scipy==1.13.0
seaborn==0.13.2
semver==2.13.0
Send2Trash==1.8.2
sentencepiece==0.2.0
sentry-sdk==1.44.1
sgmllib3k==1.0.0
shapely==2.0.1
shellingham==1.5.4
six==1.16.0
skypilot==0.5.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.29
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.37.2
streamlit==1.33.0
structlog==24.1.0
supervision==0.20.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.4.0
terminado==0.18.1
texttable==1.7.0
thop==0.1.1.post2209072238
tifffile==2024.4.24
tiktoken==0.6.0
time-machine==2.14.1
timm==0.9.16
tinycss2==1.2.1
tinydb==4.8.0
tinysegmenter==0.3
tldextract==5.1.2
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torchvision==0.17.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
transformers==4.37.2
typer
types-pkg-resources==0.1.3
types-python-dateutil==2.9.0.20240316
types-PyYAML==6.0.12.20240311
types-requests==2.31.0.6
types-urllib3==1.26.25.14
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
ua-parser==0.18.0
ujson==5.9.0
ultralytics==8.1.42
uri-template==1.3.0
uritemplate==4.1.1
urllib3==1.26.18
user-agents==2.2.0
uv==0.1.39
uvicorn==0.29.0
uvloop==0.19.0
vllm_nccl_cu12==2.18.1.0.4.0
watchdog==4.0.0
watchfiles==0.21.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==11.0.3
widgetsnbextension==4.0.10
xmljson==0.2.0
xxhash==3.4.1
yapf==0.40.2
yarl==1.9.4
zipp==3.18.1
zstandard==0.22.0
zxing-cpp==2.2.0
